fix: picker separation
