+++ agenda/presentation/src/main/java/com/juandgaines/agenda/presentation/AgendaScreen.kt
@@ -9,21 +9,18 @@ import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.size
-import androidx.compose.material3.DatePicker
-import androidx.compose.material3.DatePickerDialog
 import androidx.compose.material3.ExperimentalMaterial3Api
 import androidx.compose.material3.FabPosition
 import androidx.compose.material3.Icon
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Text
-import androidx.compose.material3.TextButton
-import androidx.compose.material3.rememberDatePickerState
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
+import com.juandgaines.agenda.componets.AgendaDatePicker
 import com.juandgaines.agenda.componets.ProfileIcon
 import com.juandgaines.agenda.componets.selector_date.DateSelector
 import com.juandgaines.core.presentation.designsystem.AddIcon
@@ -96,35 +93,8 @@ fun AgendaScreen(
         )
 
         if (stateAgenda.isDatePickerOpened) {
-            val datePickerState = rememberDatePickerState()
-            DatePickerDialog(
-               onDismissRequest = {
-                    agendaActions(AgendaActions.DismissDateDialog)
-               },
-                confirmButton = {
-                    TextButton(
-                        onClick = {
-                            datePickerState.selectedDateMillis?.let { date->
-                                agendaActions(AgendaActions.SelectDate(date))
-                            }
-                        }
-                    ){
-                        Text("Select")
-                    }
-                },
-                dismissButton = {
-                    TextButton(
-                        onClick = {
-                            agendaActions(AgendaActions.DismissDateDialog)
-                        }
-                    ) {
-                        Text("Cancel")
-                    }
-                },
-            ){
-                DatePicker(
-                    state = datePickerState
-                )
+            AgendaDatePicker { action ->
+                agendaActions(action)
             }
         }
     }
