+++ agenda/presentation/src/main/res/values/strings.xml
@@ -1,0 +1,5 @@
+<?xml version="1.0" encoding="utf-8"?>
+<resources>
+  <string name="select">Select</string>
+  <string name="cancel">Cancel</string>
+</resources>
